+++ packages/notion-client/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @julianjark/notion-client
 
+## 0.9.0
+
+### Minor Changes
+
+- fetch all entries from notion database
+
 ## 0.8.0
 
 ### Minor Changes